% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{MSfrags}
\alias{MSfrags}
\title{MSfrags}
\format{
A character vector of length 4 with peptide sequences
}
\source{
data-raw/data.R
}
\usage{
MSfrags
}
\description{
This is the sample data set that will be provided with this package, it is some protein
fragments with modifications we may expect from the CCR5 gene in the mass spec experiment
}
\keyword{datasets}
