% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.protein_mods}
\alias{print.protein_mods}
\title{print.R
Print method for protein_mods objects}
\usage{
\method{print}{protein_mods}(x, ...)
}
\arguments{
\item{x}{A \code{protein_mods} object, which is a data frame with the amino acid sequences and modifications.}

\item{...}{Other arguments passed to or from \code{print}}
}
\value{
A printed out list of protein sequences and where the mutation occurs
}
\description{
print.R
Print method for protein_mods objects
}
